refactor: split persistence functionality
